integrating updates to ml4gw and hermes main branches (#406)
* updating ml4gw and hermes references

* updating to use ml4gw utils as much as possible

* standardizing PsdLength inputs and fixing tests

* getting rid of normalize_psds in datagen

* updating architecture lib init

* getting train tests passing
