+++ libs/architectures/aframe/architectures/__init__.py
@@ -1,3 +1,3 @@
-from .preprocessor import Preprocessor
+from .preprocessor import BackgroundSnapshotter, BatchWhitener, PsdEstimator
 from .resnet import BottleneckResNet, ResNet
 from .wrapper import architecturize, get_arch_fns
