+++ projects/sandbox/datagen/datagen/utils/timeslide_waveforms.py
@@ -3,11 +3,9 @@ import time
 from pathlib import Path
 from typing import List, Tuple
 
-import h5py
 import numpy as np
 import torch
-
-from ml4gw.spectral import normalize_psd
+from gwpy.timeseries import TimeSeriesDict
 
 
 def calc_shifts_required(Tb: float, T: float, delta: float) -> int:
@@ -58,16 +56,13 @@ def io_with_blocking(f, fname, timeout=10):
                 raise
 
 
-def load_psds(
-    background: Path, ifos: List[str], sample_rate: float, df: float
-):
+def load_psds(background: Path, ifos: List[str], df: float) -> torch.Tensor:
     """Calculate PSDs from background generated by `background.py`"""
-    with h5py.File(background, "r") as f:
-        psds = []
-        for ifo in ifos:
-            hoft = f[ifo][:]
-            psd = normalize_psd(hoft, df, sample_rate)
-            psds.append(psd)
+    background = TimeSeriesDict.read(background, path=ifos)
+    psds = []
+    for ifo in ifos:
+        psd = background[ifo].psd(1 / df, window="hann", method="median")
+        psds.append(psd.value)
     psds = torch.tensor(np.stack(psds), dtype=torch.float64)
     return psds
