+++ projects/sandbox/train/tests/test_augmentor.py
@@ -5,6 +5,8 @@ import pytest
 import torch
 from train.augmentor import AframeBatchAugmentor
 
+from aframe.architectures.preprocessor import PsdEstimator
+
 
 @pytest.fixture
 def sample_rate():
@@ -37,24 +39,17 @@ def sample(obj, N, kernel_size, _):
 
 
 # dummy psd estimator and whitener
-# TODO: implement local whitener in ml4gw
 @pytest.fixture
-def psd_estimator(background_length, sample_rate):
-    def f(X):
-        size = background_length * sample_rate
-        splits = [size, X.shape[-1] - size]
-        background, X = torch.split(X, splits, dim=-1)
-        return X, background
-
-    return f
+def psd_estimator(kernel_length, fduration, sample_rate):
+    return PsdEstimator(kernel_length + fduration, sample_rate, 2)
 
 
 @pytest.fixture
 def whitener():
-    def f(x, y):
-        return x
+    def skip(X, psds):
+        return X * 2
 
-    return f
+    return skip
 
 
 rand_value = 0.1 + 0.5 * (torch.arange(32) % 2)
@@ -90,7 +85,7 @@ def test_bbhnet_batch_augmentor(
     y = torch.zeros((32, 1))
 
     X, y = augmentor(X, y)
-    assert (X[::2] == 1).all().item()
+    assert (X[::2] == 2).all().item()
     assert (X[1::2] == 0).all().item()
     assert (y[::2] == 1).all().item()
     assert (y[1::2] == 0).all().item()
@@ -98,7 +93,9 @@ def test_bbhnet_batch_augmentor(
     for _, tensor in augmentor.polarizations.items():
         assert len(tensor) == augmentor.num_waveforms
 
-    assert len(list(augmentor.buffers())) == 2
+    # 2 buffers for the ifo geometry
+    # 2 buffers belong to the whitener
+    assert len(list(augmentor.buffers())) == 4
 
     polarizations = {
         "plus": torch.randn(100, 4096),
@@ -160,14 +157,10 @@ def test_bbhnet_batch_augmentor_with_swapping_and_muting(
 
     X = torch.zeros((32, 2, size))
     y = torch.zeros((32, 1))
-    # glitch_sampler = Mock(return_value=(X, y))
-    # glitch_sampler.prob = 0.0
-
     augmentor = AframeBatchAugmentor(
         sample_rate=sample_rate,
         ifos=["H1", "L1"],
         signal_prob=0.5,
-        # glitch_sampler=glitch_sampler,
         trigger_distance=-0.25,
         dec=lambda N: torch.zeros((N,)),
         psi=lambda N: torch.zeros((N,)),
@@ -235,7 +228,6 @@ def test_sample_responses(
         ifos,
         sample_rate,
         signal_prob,
-        # glitch_sampler,
         dec=lambda N: torch.randn(N),
         psi=lambda N: torch.randn(N),
         phi=lambda N: torch.randn(N),
